Extract SwissTransferBottomSheet to its own component file
